requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pdfplumber
python-docx
sqlalchemy
httpx
numpy
jinja2
python-dotenv
pytest
psycopg2-binary
reportlab

# LangGraph for workflow orchestration
langgraph

# Google Gemini AI
google-generativeai
